+++ pyproject.toml
@@ -55,7 +55,12 @@ docs = [
     "pymdown-extensions==10.2.1",
 
 ]
-tests = ["pytest==7.4.0", "pytest-cov==4.1.0", "mp-api==0.35.1"]
+tests = [
+    "mp-api==0.35.1",
+    "pytest==7.4.0",
+    "pytest-cov==4.1.0",
+    "pytest-xdist==3.3.1",
+]
 strict = [
     "numba==0.57.0",
     "pymatgen==2023.08.10",
