+++ src/rxn_network/pathways/solver.py
@@ -104,7 +104,7 @@ class PathwaySolver(Solver):
         entries: GibbsEntrySet,
         cost_function: CostFunction,
         open_elem: str | Element | None = None,
-        chempot: float | None = None,
+        chempot: float = 0.0,
         chunk_size: int = 100000,
         batch_size: int | None = None,
     ):
@@ -113,8 +113,15 @@ class PathwaySolver(Solver):
             pathways: List of reaction pathways derived from the network.
             entries: GibbsEntrySet containing all entries in the network.
             cost_function: CostFunction object to use for the solver.
-            open_elem: Element to use for pathways with an open element.
+            open_elem: Optional element to use for pathways with an open element.
             chempot: Chemical potential to use for pathways with an open element.
+                Defaults to 0.0.
+            chunk_size: The number of pathways per chunk to use for balancing. Defaults
+                to 100,000.
+            batch_size: Number of chunks to submit to each CPU at a time. Automatically
+                calculated if not set.
+
+
         """
         super().__init__(pathways=deepcopy(pathways))
         self._entries = entries
