+++ tests/costs/test_softplus.py
@@ -34,10 +34,7 @@ def test_missing_parameter(softplus_with_attr_and_param, computed_rxn):
 
 
 def test_repr(softplus_with_attr, softplus_with_attr_and_param):
-    assert (
-        repr(softplus_with_attr)
-        == "Softplus with parameters: energy_per_atom (1.0)"
-    )
+    assert repr(softplus_with_attr) == "Softplus with parameters: energy_per_atom (1.0)"
     assert (
         repr(softplus_with_attr_and_param)
         == "Softplus with parameters: energy_per_atom (0.3) test_param (0.7)"
