+++ tests/entries/test_utils.py
